- 修复map文件夹下快捷方式无法启动问题
